update
Footer will not be showed by default, by with condition. Home button will be showed if user have already logged in. Fix color bug for slideBar. Create page when user click on each button and change the color of that button. Make the default page to About us Page.

There is the variable named userLoggenIn in Aoo.jsx for you to test Home button.